% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base.R
\docType{data}
\name{bin.data}
\alias{bin.data}
\title{An artificially constructed \link{BinData-class} class object.}
\format{a \link{BinData-class} class object.}
\description{
This data set contains a typical example for a BinData class object,.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

