\docType{class}
\name{BinData-class}
\alias{BinData-class}
\title{An S-4 class containing the model fit information for a CSSP model.}
\description{
\describe{ \item{chrID}{The chromosome ID.}
\item{coord}{The genome coordinates for the starting
positions of each bin.} \item{tagCount}{The number of ChIP
reads mapped to each bin.} \item{mappability}{The
mappability score of each bin.} \item{gcContent}{The
gc-content score of each bin.} \item{input}{The number of
input reads mapped to each bin.} \item{dataType}{Either
"unique" or "multi'.} }
}

