% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base.R
\docType{data}
\name{bindata.chr1}
\alias{bindata.chr1}
\title{An artificially constructed data.frame object that can be used by cssp.fit function.}
\format{a \link{data.frame} class object.}
\description{
This data set contains a typical example for a data.frame object that can be imported by \link{cssp.fit}.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

