\docType{data}
\name{bindpos}
\alias{bindpos}
\title{An artificially constructed dataset containing enrichment positions on 5 chromosomes.}
\format{A \link{list} containing the genome coordinates for enrichment sites on each of the 5 chromosomes.}
\usage{
example
}
\description{
This data set contains artificially generated
nucleotide-leve enrichment positions on a genome of 5
chromosomes.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

