% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base.R
\docType{data}
\name{sampleFit}
\alias{sampleFit}
\title{A "CSSPFit" class object containing the fitted CSSP model for \link{bin.data}.}
\format{a \link{CSSPFit-class} class object.}
\description{
A \link{CSSPFit-class} class object constructed by fitting CSSP model on \link{bin.data}.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

