% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base.R
\docType{data}
\name{tagdat_chip}
\alias{tagdat_chip}
\title{An artificially constructed dataset containing genome coordinates for aligned ChIP sample reads.}
\format{a \link{list} containing the reads coordinates on each of the 5 chromosomes.}
\usage{
example
}
\description{
This dataset contains artificially generated genome coordinates for ChIP sample reads on a genome of 5 chromosomes. The sign of each read represents the strand direction, with 5' represented by positive numbers and 3' represented by negative numbers.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

