\docType{data}
\name{tagdat_input}
\alias{tagdat_input}
\title{An artificially constructed dataset containing genome coordinates for aligned input sample reads.}
\format{a \link{list} containing the reads coordinates on each of the 5 chromosomes.}
\usage{
example
}
\description{
This dataset contains artificially generated genome
coordinates for ChIP sample reads on a genome of 5
chromosomes. The sign of each read represents the strand
direction, with 5' represented by positive numbers and 3'
represented by negative numbers.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

