\docType{data}
\name{peakpos}
\alias{peakpos}
\title{An artificially generated dataset containing peak invervals on 5 chromosomes.}
\format{a \link{list} of 2-column matrices. Each matrix contains the coordinates of the peak intervals for one chromosome.}
\description{
This data set contains the genome coordinates of
artificially generated peak intervals on a genome of 5
chromosomes.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

